+++ tests/alltags.js
@@ -1,5 +1,13 @@
 
-var sm = require('../dist/index')
+var { createSitemap }= require('../dist/index')
 
 var config = require('./sampleconfig.json')
-  console.log(sm.createSitemap(config).toString())
+  console.log(createSitemap(config).toString(true))
+  /*
+let urls = []
+config.urls.forEach((smi) => {
+  urls.push(Sitemap.normalizeURL(smi, undefined, 'https://roosterteeth.com'))
+})
+config.urls = urls
+  console.log(JSON.stringify(config))
+  */